初始化docker-compose配置, maven依赖, README

+++ src/main/java/com/mycompany/app/mapper/HelloMapper.java
@@ -5,5 +5,5 @@ import com.mycompany.app.domain.Hello;
 
 @Mapper
 public interface HelloMapper {
-    Hello findOneById(int id);
+    Hello findOneById(Long id);
 }
